+++ db/migrate/20170608131624_add_file_size_to_images.rb
@@ -1,0 +1,5 @@
+class AddFileSizeToImages < ActiveRecord::Migration
+  def change
+    add_column :integral_images, :file_size , :integer
+  end
+end